moved the protocol to the network module
