+++ raiden/raiden_service.py
@@ -16,7 +16,7 @@ from raiden.channelgraph import ChannelGraph
 from raiden.tasks import AlarmTask, LogListenerTask, StartExchangeTask, HealthcheckTask
 from raiden.encoding import messages
 from raiden.messages import SignedMessage
-from raiden.raiden_protocol import RaidenProtocol
+from raiden.network.protocol import RaidenProtocol
 from raiden.utils import privatekey_to_address, isaddress, pex, GLOBAL_CTX
 
 log = slogging.get_logger(__name__)  # pylint: disable=invalid-name
